requirements.txt
amqp==5.3.1
asgiref==3.11.0
celery==5.6.0
channels==4.3.2
channels-redis==4.3.0
daphne==4.2.1
Django==6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.10.1
redis==7.1.0
requests==2.32.5
stripe==14.1.0
